Move benchmark unit tests to separate files.
- move constant and fe speed tests to new files
- rename speed unit test of FieldModel
