+++ unit_tests/fields/field_const_speed_test.cpp
@@ -1,16 +1,17 @@
 /*
- * field_evaluate_model_test.cpp
+ * field_const_speed_test.cpp
  *
- *  Created on: Sept 21, 2020
+ *  Created on: Dec 03, 2019
  *      Author: David Flanderka
  *
- *  Tests evaluation of FieldModel
+ *  Speed tests of FieldConstant
  */
 
 #define TEST_USE_MPI
 #define FEAL_OVERRIDE_ASSERTS
 #include <flow_gtest_mpi.hh>
 #include <mesh_constructor.hh>
+#include "arma_expect.hh"
 
 #include "fields/eval_points.hh"
 #include "fields/eval_subset.hh"
@@ -18,8 +19,6 @@
 #include "fields/field_values.hh"
 #include "fields/field_set.hh"
 #include "tools/unit_si.hh"
-#include "fields/field.hh"
-#include "fields/field_model.hh"
 #include "quadrature/quadrature.hh"
 #include "quadrature/quadrature_lib.hh"
 #include "fem/dofhandler.hh"
@@ -37,19 +36,16 @@ class Balance;
 
 
 /****************************************************************************************
- *                 Speed test of FieldModel evaluation
+ *                 Speed test of FieldConstant evaluation
  *
  * Results:
  * Mesh 27936 elements, 100 assemblation loops
  * Checked GenericAssembly with active bulk integral only vs. with all active integrals
- * (times in brackets are )
  *
- *                           bulk                     all
- * add_integrals_to_patch   18.99                    42.38
- * create_patch              2.97                    18.30
- * cache_update             10.30  11.63* -8.43**    31.38  34.19* -26.38**
- * *) times of cache_update with functions passed to model ( not struct and operator() )
- * **) times of FieldConstant fields of field_evaluate_const_test
+ *                           bulk      all
+ * add_integrals_to_patch   19.12    43.95
+ * create_patch              3.00    18.59
+ * cache_update              8.43    26.38
  *
  ****************************************************************************************/
 
@@ -57,92 +53,47 @@ class Balance;
 
 static const unsigned int profiler_loop = 100;
 
-using Sclr = double;
-using Vect = arma::vec3;
-using Tens = arma::mat33;
-
-// Functors of FieldModels
-struct fn_model_scalar {
-    inline Sclr operator() (Vect vec) {
-        return arma::norm(vec, 2);
-    }
-};
-
-struct fn_model_vector {
-	inline Vect operator() (Sclr scal, Vect vec) {
-        return scal * vec + vec;
-    }
-};
-
-struct fn_model_tensor {
-	inline Tens operator() (Sclr scal, Tens ten) {
-        return 0.5 * scal * ten;
-    }
-};
+// allow running tests of all type of fields (unset allow runnig only vector field)
+#define ALL_FIELDS
 
 
-class FieldModelSpeedTest : public testing::Test {
+class FieldConstantSpeedTest : public testing::Test {
 public:
     class EqData : public FieldSet {
     public:
         EqData() : order(2) {
-            *this += vector_const_field
-                        .name("vector_const_field")
-                        .description("Constant vector.")
-                        .input_default("0.0")
-                        .flags_add(in_main_matrix)
-                        .units( UnitSI().kg(3).m() );
-            *this += scalar_const_field
-                        .name("scalar_const_field")
-                        .description("Constant scalar")
-                        .units( UnitSI().m() );
-            *this += tensor_const_field
-                        .name("tensor_const_field")
-                        .description("Constant tensor")
-                        .units( UnitSI::dimensionless() )
-                        .flags_add(in_main_matrix);
             *this += vector_field
                         .name("vector_field")
                         .description("Velocity vector.")
                         .input_default("0.0")
                         .flags_add(in_main_matrix)
                         .units( UnitSI().kg(3).m() );
+#ifdef ALL_FIELDS
             *this += scalar_field
                         .name("scalar_field")
                         .description("Pressure head")
-                        .input_default("0.0")
                         .units( UnitSI().m() );
             *this += tensor_field
                         .name("tensor_field")
                         .description("")
-                        .input_default("0.0")
                         .units( UnitSI::dimensionless() )
                         .flags_add(in_main_matrix);
+#endif // ALL_FIELDS
 
-        }
-
-        /// Initialiye FieldModels
-        void initialize() {
-            scalar_field.set(Model<3, FieldValue<3>::Scalar>::create(fn_model_scalar(), vector_const_field), 0.0);
-            vector_field.set(Model<3, FieldValue<3>::VectorFixed>::create(fn_model_vector(), scalar_const_field, vector_const_field), 0.0);
-            tensor_field.set(Model<3, FieldValue<3>::TensorFixed>::create(fn_model_tensor(), scalar_const_field, tensor_const_field), 0.0);
         }
 
     	/// Polynomial order of finite elements.
     	unsigned int order;
 
-    	// constant fields, we need these fields to create models
-        Field<3, FieldValue<3>::Scalar > scalar_const_field;
-        Field<3, FieldValue<3>::VectorFixed > vector_const_field;
-        Field<3, FieldValue<3>::TensorFixed > tensor_const_field;
-
-        // computing fields of FieldModels
-        Field<3, FieldValue<3>::Scalar > scalar_field;
+    	// fields
         Field<3, FieldValue<3>::VectorFixed > vector_field;
+#ifdef ALL_FIELDS
+        Field<3, FieldValue<3>::Scalar > scalar_field;
         Field<3, FieldValue<3>::TensorFixed > tensor_field;
+#endif // ALL_FIELDS
     };
 
-    FieldModelSpeedTest() : tg_(0.0, 1.0) {
+    FieldConstantSpeedTest() : tg_(0.0, 1.0) {
     	Profiler::instance();
 
         FilePath::set_io_dirs(".",UNIT_TESTS_SRC_DIR,"",".");
@@ -150,11 +101,12 @@ public:
         PetscInitialize(0,PETSC_NULL,PETSC_NULL,PETSC_NULL);
 
         data_ = std::make_shared<EqData>();
+        data_->add_coords_field();
         mesh_ = mesh_full_constructor("{ mesh_file=\"mesh/test_27936_elem.msh\", optimize_mesh=false }");
         dh_ = std::make_shared<DOFHandlerMultiDim>(*mesh_);
     }
 
-    ~FieldModelSpeedTest() {
+    ~FieldConstantSpeedTest() {
         Profiler::uninitialize();
     }
 
@@ -162,13 +114,12 @@ public:
         return IT::Record("SomeEquation","")
                 .declare_key("data", IT::Array(
                         IT::Record("SomeEquation_Data", FieldCommon::field_descriptor_record_description("SomeEquation_Data") )
-                        .copy_keys( FieldModelSpeedTest::EqData().make_field_descriptor_type("SomeEquation") )
-                        .declare_key("scalar_const_field", FieldAlgorithmBase< 3, FieldValue<3>::Scalar >::get_input_type_instance(), "" )
-                        .declare_key("vector_const_field", FieldAlgorithmBase< 3, FieldValue<3>::VectorFixed >::get_input_type_instance(), "" )
-                        .declare_key("tensor_const_field", FieldAlgorithmBase< 3, FieldValue<3>::TensorFixed >::get_input_type_instance(), "" )
-//                        .declare_key("scalar_field", FieldAlgorithmBase< 3, FieldValue<3>::Scalar >::get_input_type_instance(), "" )
-//                        .declare_key("vector_field", FieldAlgorithmBase< 3, FieldValue<3>::VectorFixed >::get_input_type_instance(), "" )
-//                        .declare_key("tensor_field", FieldAlgorithmBase< 3, FieldValue<3>::TensorFixed >::get_input_type_instance(), "" )
+                        .copy_keys( FieldEvalConstantTest::EqData().make_field_descriptor_type("SomeEquation") )
+                        .declare_key("vector_field", FieldAlgorithmBase< 3, FieldValue<3>::VectorFixed >::get_input_type_instance(), "" )
+#ifdef ALL_FIELDS
+                        .declare_key("scalar_field", FieldAlgorithmBase< 3, FieldValue<3>::Scalar >::get_input_type_instance(), "" )
+                        .declare_key("tensor_field", FieldAlgorithmBase< 3, FieldValue<3>::TensorFixed >::get_input_type_instance(), "" )
+#endif // ALL_FIELDS
                         .close()
                         ), IT::Default::obligatory(), ""  )
                 .close();
@@ -187,13 +138,12 @@ public:
 
         data_->set_mesh(*mesh_);
         data_->set_input_list( inputs[input_last], tg_ );
-        data_->initialize();
         data_->set_time(tg_.step(), LimitSide::right);
     }
 
 
 	void profiler_output() {
-		static ofstream os( FilePath("speed_eval_model_test.log", FilePath::output_file) );
+		static ofstream os( FilePath("speed_eval_const_test.log", FilePath::output_file) );
 		Profiler::instance()->output(MPI_COMM_WORLD, os);
 		os << "" << std::setfill('=') << setw(80) << "" << std::setfill(' ') << endl << endl;
 	}
@@ -207,10 +157,10 @@ public:
 template <unsigned int dim>
 class AssemblyDimTest : public AssemblyBase<dim> {
 public:
-    typedef typename FieldModelSpeedTest::EqData EqFields;
-    typedef typename FieldModelSpeedTest::EqData EqData;
+    typedef typename FieldConstantSpeedTest::EqData EqFields;
+    typedef typename FieldConstantSpeedTest::EqData EqData;
 
-    static constexpr const char * name() { return "AssemblyModelTest"; }
+    static constexpr const char * name() { return "AssemblyConstantTest"; }
 
     /// Constructor.
     AssemblyDimTest(EqFields *eq_fields, EqData *eq_data)
@@ -236,14 +186,14 @@ string eq_data_input_speed = R"YAML(
 data:
   - region: ALL
     time: 0.0
-    scalar_const_field: !FieldConstant
+    scalar_field: !FieldConstant
       value: 0.5
-    vector_const_field: [1, 2, 3]
-    tensor_const_field: [0.1, 0.2, 0.3, 0.4, 0.5, 0.6]
+    vector_field: [1, 2, 3]
+    tensor_field: [0.1, 0.2, 0.3, 0.4, 0.5, 0.6]
 )YAML";
 
 
-TEST_F(FieldModelSpeedTest, speed_test) {
+TEST_F(FieldConstantSpeedTest, speed_test) {
 	this->read_input(eq_data_input_speed);
 
 	GenericAssembly< AssemblyDimTest > ga_bulk(data_.get(), data_.get());
